+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutStatusView.java
@@ -24,12 +24,8 @@ import io.snabble.sdk.ui.scanner.BarcodeView;
 import io.snabble.sdk.ui.telemetry.Telemetry;
 
 class CheckoutStatusView extends FrameLayout implements Checkout.OnCheckoutStateChangedListener {
-    private ViewGroup[] steps;
-    private int currentStep;
     private Checkout checkout;
     private BarcodeView checkoutIdCode;
-
-    private Handler handler = new Handler(Looper.getMainLooper());
 
     public CheckoutStatusView(Context context) {
         super(context);
@@ -48,12 +44,6 @@ class CheckoutStatusView extends FrameLayout implements Checkout.OnCheckoutState
 
     private void inflateView() {
         inflate(getContext(), R.layout.view_checkout_status, this);
-        currentStep = 0;
-
-        steps = new ViewGroup[3];
-        steps[0] = findViewById(R.id.step1);
-        steps[1] = findViewById(R.id.step2);
-        steps[2] = findViewById(R.id.step3);
 
         checkoutIdCode = findViewById(R.id.checkout_id_code);
 
@@ -71,66 +61,12 @@ class CheckoutStatusView extends FrameLayout implements Checkout.OnCheckoutState
             }
         });
 
-        for(View v : steps) {
-            ImageView imageView = v.findViewWithTag("image");
-            ProgressBar progressBar = v.findViewWithTag("progress");
-
-            int color = 0x1d1f2440;
-            progressBar.getIndeterminateDrawable().setColorFilter(color, PorterDuff.Mode.SRC_IN);
-            progressBar.getProgressDrawable().setColorFilter(color, PorterDuff.Mode.SRC_IN);
-            progressBar.setIndeterminate(false);
-            progressBar.setProgress(0);
-
-            // tinting the white backgrounded images with the background color
-            // so they appear seamless
-            ImageViewCompat.setImageTintMode(imageView, PorterDuff.Mode.MULTIPLY);
-            TypedValue a = new TypedValue();
-            getContext().getTheme().resolveAttribute(android.R.attr.windowBackground, a, true);
-            if (a.type >= TypedValue.TYPE_FIRST_COLOR_INT && a.type <= TypedValue.TYPE_LAST_COLOR_INT) {
-                ImageViewCompat.setImageTintList(imageView, ColorStateList.valueOf(a.data));
-            }
-        }
-
         if (SnabbleUI.getActionBar() != null) {
-            SnabbleUI.getActionBar().setTitle(R.string.Snabble_Checkout_verifying);
+            SnabbleUI.getActionBar().setTitle(R.string.Snabble_Payment_confirm);
         }
 
         checkout = SnabbleUI.getProject().getCheckout();
         onStateChanged(checkout.getState());
-    }
-
-    private void setStep(int step) {
-        currentStep = step;
-
-        if (currentStep < steps.length) {
-            crossFade(steps[currentStep]);
-
-            int nextStep = currentStep + 1;
-            if (nextStep < steps.length) {
-                ProgressBar progressBar = steps[nextStep].findViewWithTag("progress");
-                progressBar.setIndeterminate(true);
-            }
-        }
-    }
-
-    private void crossFade(final ViewGroup vg) {
-        ImageView imageView = vg.findViewWithTag("image");
-        ImageView imageView2 = vg.findViewWithTag("image_ok");
-        ProgressBar progressBar = vg.findViewWithTag("progress");
-
-        imageView.animate()
-                .alpha(0f)
-                .setInterpolator(new AccelerateDecelerateInterpolator())
-                .setDuration(500)
-                .start();
-
-        imageView2.animate()
-                .alpha(1f)
-                .setInterpolator(new AccelerateDecelerateInterpolator())
-                .setDuration(500)
-                .start();
-
-        progressBar.setVisibility(View.INVISIBLE);
     }
 
     @Override
@@ -159,17 +95,6 @@ class CheckoutStatusView extends FrameLayout implements Checkout.OnCheckoutState
                 if (id != null) {
                     checkoutIdCode.setText(id);
                 }
-                setStep(0);
-                break;
-            case PAYMENT_APPROVED:
-                setStep(1);
-                Telemetry.event(Telemetry.Event.CheckoutSuccessful);
-                handler.postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        setStep(2);
-                    }
-                }, 1000);
                 break;
             case DENIED_BY_PAYMENT_PROVIDER:
                 Telemetry.event(Telemetry.Event.CheckoutDeniedByPaymentProvider);